refactor: constants directory
