--- constants.ts
@@ -1,25 +1,0 @@
-export const VALUE_TYPE = {
-  NUMBER: 29,
-  LABEL: 31,
-  TIME: 32,
-  PERCENTAGE: 30,
-  DATE: 28,
-  COLOR: 26,
-  PIXELS: 19,
-  BOOLEAN: 2,
-  NULL: 1,
-  UNDEFINED: 0,
-} as const;
-
-export const VALUE_LIMITS = {
-  NUMBER_MAX: 1000000000,
-  NUMBER_MIN: -100000000000,
-  TIMESTAMP_MAX: 86400000,
-  TIMESTAMP_MIN: -86400000,
-
-  PERCENTAGE_100: 100000000,
-  PERCENTAGE_0: 0,
-  PERCENTAGE_MINUS_100: -100000000,
-} as const;
-
-export const BROADCAST_ID = 255;
